Convert dt to double

+++ P2-Unscented-Kalman-Filter/CarND-Unscented-Kalman-Filter-Project/src/ukf.cpp
@@ -143,7 +143,7 @@ void UKF::ProcessMeasurement(MeasurementPackage meas_package) {
 		return;
 	}
 
-	float dt = (meas_package.timestamp_ - previous_timestamp_) / 1000000.0;
+	double dt = (meas_package.timestamp_ - previous_timestamp_) / 1000000.0;
 	previous_timestamp_ = meas_package.timestamp_;
 	if(dt > 0.001) {
 		Prediction(dt);
